requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
greenlet==3.0.3
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
protobuf==3.20.3
SQLAlchemy==2.0.29
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
Flask-WTF==1.0.0